requirements.txt
cffi==1.16.0
cryptography==42.0.5
mpmath==1.3.0
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
pycparser==2.22
PyWavelets==1.6.0
reedsolo==1.7.0
sympy==1.12